feat: add webworker and resolve timeout issue on loading images
