+++ frontend/src/features/Dashboard/components/Pdf/v2/Table/index.tsx
@@ -3,7 +3,7 @@ import { StyleSheet, Text, View, Image } from '@react-pdf/renderer'
 
 import { layoutStyle } from '../style'
 
-import type { ExportImageType } from '../../../Layers/ExportLayer'
+import type { ExportImageType } from '../../../../hooks/useExportImages'
 import type { VigilanceArea } from '@features/VigilanceArea/types'
 import type { AMPFromAPI } from 'domain/entities/AMPs'
 import type { RegulatoryLayerWithMetadata } from 'domain/entities/regulatory'
