+++ frontend/src/features/Dashboard/components/Pdf/v2/utils.ts
@@ -1,6 +1,6 @@
 import { Dashboard } from '@features/Dashboard/types'
 
-import type { ExportImageType } from '../../Layers/ExportLayer'
+import type { ExportImageType } from '../../../hooks/useExportImages'
 
 export function getImage(images: ExportImageType[], type: Dashboard.Layer, id: number | undefined): string | undefined {
   return images.find(image => {
