+++ frontend/src/features/Dashboard/utils.tsx
@@ -2,10 +2,19 @@ import { ampsAPI } from '@api/ampsAPI'
 import { regulatoryLayersAPI } from '@api/regulatoryLayersAPI'
 import { reportingsAPI } from '@api/reportingsAPI'
 import { vigilanceAreasAPI } from '@api/vigilanceAreasAPI'
+import { getAMPFeature } from '@features/map/layers/AMP/AMPGeometryHelpers'
+import { getRegulatoryFeature } from '@features/map/layers/Regulatory/regulatoryGeometryHelpers'
+import { getVigilanceAreaZoneFeature } from '@features/VigilanceArea/components/VigilanceAreaLayer/vigilanceAreaGeometryHelper'
 import { isCypress } from '@utils/isCypress'
 
-import type { Dashboard } from './types'
+import { Dashboard } from './types'
+
+import type { VigilanceArea } from '@features/VigilanceArea/types'
+import type { EntityState } from '@reduxjs/toolkit'
 import type { HomeRootState } from '@store/index'
+import type { AMP } from 'domain/entities/AMPs'
+import type { RegulatoryLayerWithMetadata } from 'domain/entities/regulatory'
+import type { Feature } from 'ol'
 import type { Action } from 'redux'
 import type { ThunkDispatch } from 'redux-thunk'
 
@@ -41,4 +50,62 @@ export async function populateExtractAreaFromApi(
       extractedAreaFromApi.vigilanceAreaIds.includes(vigilanceArea.id)
     )
   }
+}
+
+export const extractFeatures = (
+  dashboard: Dashboard.Dashboard | undefined,
+  regulatoryLayers: EntityState<RegulatoryLayerWithMetadata, number> | undefined,
+  ampLayers: EntityState<AMP, number> | undefined,
+  vigilanceAreas: EntityState<VigilanceArea.VigilanceAreaLayer, number> | undefined
+) => {
+  const allFeatures: Feature[] = []
+
+  // Récupération des zones réglementaires
+  if (dashboard?.regulatoryAreaIds) {
+    dashboard.regulatoryAreaIds.forEach(layerId => {
+      const layer = regulatoryLayers?.entities[layerId]
+      if (layer?.geom?.coordinates.length) {
+        const feature = getRegulatoryFeature({
+          code: Dashboard.featuresCode.DASHBOARD_REGULATORY_AREAS,
+          isolatedLayer: undefined,
+          layer
+        })
+        if (feature) {
+          allFeatures.push(feature)
+        }
+      }
+    })
+  }
+
+  // AMP Features
+  if (dashboard?.ampIds) {
+    dashboard.ampIds.forEach(layerId => {
+      const layer = ampLayers?.entities[layerId]
+      if (layer?.geom?.coordinates.length) {
+        const feature = getAMPFeature({
+          code: Dashboard.featuresCode.DASHBOARD_AMP,
+          isolatedLayer: undefined,
+          layer
+        })
+        if (feature) {
+          allFeatures.push(feature)
+        }
+      }
+    })
+  }
+
+  // Zones de vigilance
+  if (dashboard?.vigilanceAreaIds) {
+    dashboard.vigilanceAreaIds.forEach(layerId => {
+      const layer = vigilanceAreas?.entities[layerId]
+      if (layer?.geom?.coordinates.length) {
+        const feature = getVigilanceAreaZoneFeature(layer, Dashboard.featuresCode.DASHBOARD_VIGILANCE_AREAS, undefined)
+        if (feature) {
+          allFeatures.push(feature)
+        }
+      }
+    })
+  }
+
+  return allFeatures
 }
